requirements.txt
python-telegram-bot==21.3
anthropic>=0.24.0
python-dotenv>=1.0.0
